// sor_refact/types/htFarmaci.ts
export type FarmaciItem = {
  codice: number
  nompaz: string
  dnasc: string
  anni: number
  sesso: string
  nureri: number
  regime: string
  annsdo: number
  numsdo: number
  codser: string
  nomser: string
  datric: string
  oraric: string
  repcam: string
  codcam: string
  numlet: number
  codmag: number
  coddep: number
  farpaz: string
  nurevt: number
  numrig: number
  nurevs: number
  nuremm: number
  nuremp: number
  nurkap: number
  codpre: number
  codart: string
  desart: string
  unimis: string
  testsomm: string
  tipsom: string
  datsom: string
  orasom: string
  qtasom: number
  qtacon: string
  qtatot: number
  numlot: string
  nummat: string
  datlot: string
  annlot: number
  meslot: number
  desnot: string
  ab: string
  note: string
}